reorganized bundle tests
