+++ tests/test_bundles.py
@@ -22,7 +22,7 @@ import pandas as pd
 from datetime import datetime
 
 #from fhir.resources.bundle import Bundle
-from py_mimic_fhir.bundle import Bundle, Bundler, rerun_bundle_from_file
+from py_mimic_fhir.bundle import Bundle, Bundler, rerun_bundle_from_file, get_n_patient_id
 
 FHIR_SERVER = os.getenv('FHIR_SERVER')
 FHIR_BUNDLE_ERROR_PATH = os.getenv('FHIR_BUNDLE_ERROR_PATH')
@@ -47,15 +47,6 @@ def get_pat_id_with_links(db_conn, resource_list):
     return resource.fhir[0]['id']
 
 
-# Get n patient_ids
-def get_n_patient_id(db_conn, n_patient):
-    q_resource = f"SELECT * FROM mimic_fhir.patient LIMIT {n_patient}"
-    resource = pd.read_sql_query(q_resource, db_conn)
-    patient_ids = [fhir['id'] for fhir in resource.fhir]
-
-    return patient_ids
-
-
 # ---------------- Test Functions --------------------
 
 
@@ -63,62 +54,6 @@ def test_bundler(db_conn):
     patient_id = get_n_patient_id(db_conn, 1)[0]
     bundler = Bundler(patient_id, db_conn)
     assert True
-
-
-def test_bundle_patient_all_resources(db_conn):
-    # Get n patient ids to then bundle and post
-    patient_id = get_n_patient_id(db_conn, 1)[0]
-    split_flag = True  # flag for breaking up bundles to smaller chunks
-
-    # Create bundle and post it
-    bundler = Bundler(patient_id)
-    bundler.generate_all_bundles()
-    response_list = bundler.post_all_bundles(
-        FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH
-    )
-
-    result = True
-    if False in response_list:
-        result = False
-    assert result
-
-
-def test_bundle_all_patient_bundles(db_conn):
-    patient_ids = get_n_patient_id(db_conn, 1)
-    split_flag = True  # Flag to subdivide bundles to speed up posting
-
-    # Create bundle and post it
-    result = True
-    for patient_id in patient_ids:
-        print(patient_id)
-        bundler = Bundler(patient_id, db_conn)
-        bundler.generate_patient_bundle()
-        response = bundler.patient_bundle.request(
-            FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH
-        )
-
-        if response == False:
-            result = False
-    assert result
-
-
-def test_bundle_multiple_patient_all_resources(db_conn):
-    # Get n patient ids to then bundle and post
-    patient_ids = get_n_patient_id(db_conn, 10)
-    split_flag = True  # Flag to subdivide bundles to speed up posting
-
-    # Create bundle and post it
-    result = True
-    for patient_id in patient_ids:
-        bundler = Bundler(patient_id, db_conn)
-        bundler.generate_all_bundles()
-        response_list = bundler.post_all_bundles(
-            FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH
-        )
-
-        if False in response_list:
-            result = False
-    assert result
 
 
 def test_bad_bundle(db_conn):
@@ -133,30 +68,6 @@ def test_bad_bundle(db_conn):
     bundle.add_entry([resource])
     response = bundle.request(FHIR_SERVER, err_path=FHIR_BUNDLE_ERROR_PATH)
     assert response == False
-
-
-def test_rerun_bundle(db_conn):
-    day_of_week = datetime.now().strftime('%A').lower()
-    err_file = f'{FHIR_BUNDLE_ERROR_PATH}err-bundles-{day_of_week}.json'
-    resp = rerun_bundle_from_file(err_file, db_conn, FHIR_SERVER)
-
-    # If only test_bad_bundle has been run then this should pass, if other issues run then it will fail
-    # If other tests have failed and written to the log, this will fail since the root causes won't be solved
-    assert resp == True
-
-
-def test_bundle_resources_from_list(db_conn):
-    q_resource = f"""
-        SELECT fhir FROM mimic_fhir.patient LIMIT 10
-    """
-    pd_resources = pd.read_sql_query(q_resource, db_conn)
-    resources = pd_resources.fhir.to_list()
-
-    bundle = Bundle()
-    bundle.add_entry(resources)
-    response = bundle.request(FHIR_SERVER)
-
-    assert response
 
 
 def test_patient_bundle(db_conn):
@@ -401,65 +312,3 @@ def test_icu_observation_bundle(db_conn):
     )
     logging.error(patient_id)
     assert response
-
-
-def test_post_100_resources(db_conn):
-    q_resource = f"""
-        SELECT fhir FROM mimic_fhir.patient LIMIT 100
-    """
-    pd_resources = pd.read_sql_query(q_resource, db_conn)
-    resources = pd_resources.fhir.to_list()
-    split_flag = True  # Divide up bundles into smaller chunks
-
-    bundle = Bundle()
-    bundle.add_entry(resources)
-    resp = bundle.request(
-        FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH, bundle_size=50
-    )
-    assert resp
-
-
-def test_bundle_multiple_lab_resources(db_conn):
-    # Get n patient ids to then bundle and post
-    patient_ids = get_n_patient_id(db_conn, 2)
-    split_flag = True  # Flag to subdivide bundles to speed up posting
-
-    # Create bundle and post it
-    result = True
-    for patient_id in patient_ids:
-        bundler = Bundler(patient_id, db_conn)
-        bundler.generate_lab_bundle()
-        resp = bundler.lab_bundle.request(
-            FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH
-        )
-
-        if resp == False:
-            result = False
-    assert result
-
-
-def test_largest_bundle():
-    # Bundle is 44,277 resources!
-    # Ran for ~20 minutes without finishing...
-    # I was posting 1000 resources in 6 seconds, so ~44,000 should be about 5 minutes...
-    # Keep playing with this
-    patient_id = '77e10fd0-6a1c-5547-a130-fae1341acf36'
-    bundler = Bundler(patient_id, db_conn)
-    bundler.generate_icu_obs_bundle()
-    split_flag = True  # send bundle in smaller chunks
-    resp = bundler.icu_obs_bundle.request(
-        FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH
-    )
-    assert resp
-
-
-def test_largest_icu_ce_bundle():
-    # Bundle is ~42,000
-    patient_id = '77e10fd0-6a1c-5547-a130-fae1341acf36'
-    bundler = Bundler(patient_id, db_conn)
-    bundler.generate_icu_ce_bundle()
-    split_flag = True  # send bundle in smaller chunks
-    resp = bundler.icu_ce_bundle.request(
-        FHIR_SERVER, split_flag, FHIR_BUNDLE_ERROR_PATH
-    )
-    assert resp
